Separated server into individual routes
Added a routes folder to separate all of the routes for the server
into their own files. Made some minor styling changes, such as removing
the button to get data since the data is now fetched on page load.
Added a header for the table to specify the names of each of the
properties.