Implemented basic error handling
Handled errors that occur when the Zendesk API is unavailable and when
there is an issue retrieving the data stored in a user's session. The
error is handled by displaying an alert with a descriptive error message
as the text. Additionally, the morgan middleware was added in order to
display all requests made to the server.